+++ worlds/turok2/docs/changelog.md
@@ -24,11 +24,12 @@ This includes changes in both the mod and AP world, in reverse order of releases
       - Scale to owned unique weapons
       - No logic (any enemy can be anywhere)
     - Enemy spawners can be randomized too
-      - Includes graveyards, Sisters of Despair (who spawn the undead), wasp nests, spiders, and hives
+      - Includes undead spawners, Sisters of Despair (who spawn the undead), wasp nests, spiders, and hives
       - Can be set to the same setting as the enemizer, or a more managable set of easy-to-kill enemies
   - Enemy trap pools are now configurable to the same pool options as the enemizer
     - They contain the full suite of enemies now as well!
 - Bug fixes
+  - Fixed random ammo packs sometimes coming back after being collected
   - Fixed dup health/ammo spawners when saving and loading from the hub
   - Fixed the Ion Capacitor "important" indicator not displaying when mission items are not randomized
   - Misc location name fixes